requirements.txt
tgcrypto
asyncio
ffmpeg-python
py-tgcalls
pyrogram==1.4.15
yt-dlp
youtube-search-python
youtube-search
python-dotenv
dnspython
gitpython
asgiref
aiofiles
aiohttp
requests
pillow
heroku3
motor
psutil
future
wget
pyyaml
telethon
speedtest-cli
deezer-python
spotipy
pymongo
requests
dnspython


